Refactor library into modules
